+++ packages/lib/src/instrumentation.ts
@@ -8,8 +8,12 @@ import {
   MeterProvider,
   MetricReader,
   PeriodicExportingMetricReader,
+  Aggregation,
+  View,
+  ExplicitBucketHistogramAggregation,
 } from "@opentelemetry/sdk-metrics";
 import { buildInfo, BuildInfo, recordBuildInfo } from "./buildInfo";
+import { HISTOGRAM_NAME } from "./constants"
 
 let IS_EAGERLY_PUSHED = false;
 let _pushMetrics = () => {};
@@ -50,10 +54,16 @@ export function init(options: initOptions) {
   exporter = options.exporter;
   // if a pushGateway is added we overwrite the exporter
   if (options.pushGateway) {
+    // TODO - use different bucket sizes
+    //     DEFAULT_BUCKETS = (.005, .01, .025, .05, .075, .1, .25, .5, .75, 1.0, 2.5, 5.0, 7.5, 10.0, INF)
+    const inMemoryExporter = new InMemoryMetricExporter(
+      AggregationTemporality.DELTA,
+    );
+
     exporter = new PeriodicExportingMetricReader({
       // 0 - using delta aggregation temporality setting
       // to ensure data submitted to the gateway is accurate
-      exporter: new InMemoryMetricExporter(AggregationTemporality.DELTA),
+      exporter: inMemoryExporter,
     });
     // Make sure the provider is initialized and exporter is registered
     getMetricsProvider();
@@ -66,7 +76,6 @@ export function init(options: initOptions) {
     } else if (interval === 0) {
       logger("Configuring autometrics to push metrics eagerly");
       IS_EAGERLY_PUSHED = true;
-      _pushMetrics();
     }
   }
 
@@ -95,19 +104,55 @@ export function eagerlyPushMetricsIfConfigured() {
 // TODO - add a way to stop the push interval
 // TODO - handle errors
 // TODO - allow configuration of timeout
+// TODO - handle case when `fetch` is undefined
+// TODO - allow custom fetch function to be passed in
 async function pushToGateway(gateway: string) {
   const exporterResponse = await exporter.collect();
+  console.log(
+    "\n\n",
+    require("util").inspect(exporterResponse, { depth: null }),
+    "\n\n",
+  );
   const serialized = new PrometheusSerializer().serialize(
     exporterResponse.resourceMetrics,
   );
 
-  await fetch(gateway, {
-    method: "POST",
-    mode: "cors",
-    body: serialized,
-  });
-  // we flush the metrics at the end of the submission to ensure the data is not repeated
-  await exporter.forceFlush();
+  if (typeof fetch === "undefined") {
+    logger("Fetch is undefined, cannot push metrics to gateway");
+    return;
+  }
+  console.log("\n\n", serialized, "\n\n");
+
+  try {
+    const response = await fetch(gateway, {
+      method: "POST",
+      mode: "cors",
+      body: serialized,
+    });
+    if (response.ok) {
+      logger("OK response from gateway");
+    } else {
+      logger(`Error pushing metrics to gateway: ${response.statusText}`);
+      logger(JSON.stringify(await response.text(), null, 2));
+    }
+  } catch (fetchError) {
+    logger(
+      `Error pushing metrics to gateway: ${
+        fetchError?.message ?? "<no error message found>"
+      }`,
+    );
+  }
+
+  try {
+    // we flush the metrics at the end of the submission to ensure the data is not repeated
+    await exporter.forceFlush();
+  } catch (error) {
+    logger(
+      `Error flushing metrics after push: ${
+        error?.message ?? "<no error message found>"
+      }`,
+    );
+  }
 }
 
 /**
@@ -123,7 +168,18 @@ export function getMetricsProvider() {
       exporter = new PrometheusExporter();
     }
 
-    autometricsMeterProvider = new MeterProvider();
+    autometricsMeterProvider = new MeterProvider({
+      views: [
+        new View({
+          aggregation: new ExplicitBucketHistogramAggregation([
+            0, 0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1, 2.5, 5,
+            7.5, 10,
+          ]),
+          instrumentName: HISTOGRAM_NAME,
+        }),
+      ],
+    });
+
     autometricsMeterProvider.addMetricReader(exporter);
   }